ensure jma and meteofrance domains do not overlap runs

+++ Sources/App/Helper/Download/Curl.swift
@@ -41,7 +41,7 @@ final class Curl {
     
     let client: HTTPClient
 
-    public init(logger: Logger, client: HTTPClient, deadLineHours: Int = 3, readTimeout: Int = 5*60, retryError4xx: Bool = true, waitAfterLastModified: TimeInterval? = nil) {
+    public init(logger: Logger, client: HTTPClient, deadLineHours: Double = 3, readTimeout: Int = 5*60, retryError4xx: Bool = true, waitAfterLastModified: TimeInterval? = nil) {
         self.logger = logger
         self.deadline = Date().addingTimeInterval(TimeInterval(deadLineHours * 3600))
         self.retryError4xx = retryError4xx
